\name{hammer.eset}
\alias{hammer.eset}
\docType{data}
\title{
RNA-seq data from Hammer, P. et al., 2010
}
\description{
RNA-seq data structured as an expressionSet, from "mRNA-seq with 
agnostic splice site discovery for nervous system transcriptomics tested in 
chronic pain" by Hammer, P. et al. (Genome Res. 2010, 20(6):847-860), 
\url{http://dx.doi.org/10.1101/gr.101204.109}.
}
\usage{
data(hammer.eset)
}
\value{
RNA-seq data structured as an expressionSet.
}
\author{
Ran Bi \email{biran@iastate.edu}, Peng Liu \email{pliu@iastate.edu}
}
\keyword{
datasets
}