\name{hammer.eset}
\alias{hammer.eset}
\docType{data}
\title{
  RNAseq data from Hammer et al., Genome Research 2010
}
\description{
  RNAseq data structured as an expressionSet, from "mRNA-seq with agnostic splice site discovery for nervous system transcriptomics tested in chronic pain" by Paul Hammer and colleagues (Genome Res. 2010. 20: 847-860), \url{http://dx.doi.org/10.1101/gr.101204.109}.
}
\usage{data(hammer.eset)}
\examples{
data(hammer.eset)
phenoData(hammer.eset)
}
\keyword{datasets}
